requirements.txt
pygithub
pygit2
aiofiles
aiohttp
aiohttp-devtools
asynctest
flake8
ipython
isort
mypy
pydantic
pytest
python-engineio
requests
sockjs
types-aiofiles
watchgod
pyyaml
types-pyyaml
